% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USCensusGeocodeR.R
\name{geocoding_api}
\alias{geocoding_api}
\title{USCensusGeocodeR}
\usage{
geocoding_api(street, city = "", state = "", zip = "")
}
\arguments{
\item{street}{This is a string or vector of strings containing the street address.}

\item{city}{This is a string or vector of strings that contain the city.}

\item{state}{This is a string or vector of strings that contain the two-letter state abbreviations.}

\item{zip}{This is a string or vector of strings that contain the 5-character zip code.}
}
\value{
dataframe
}
\description{
This package allows for individuals to easily geocode address parts using the US Census Geocoding API Services. By passing in the Street Address, City, State, and/or Zip Code, you   can find information such as the GEOID of the block group the address resides in as well as the coordinates of the block group the address resides in.
}
\examples{
geocoding_api("1600 Pennsylvania Ave","Washington","DC","22052")

}
